refactor: convert uint256 to uint128 for storage optimization

+++ test/forge/BaseTest.sol
@@ -493,7 +493,7 @@ contract BaseTest is Test {
     function _getRepaymentDetails(Id _id, address borrower)
         internal
         view
-        returns (uint128 cycleId, uint128 amountDue, uint256 endingBalance, RepaymentStatus status)
+        returns (uint128 cycleId, uint128 amountDue, uint128 endingBalance, RepaymentStatus status)
     {
         (cycleId, amountDue, endingBalance) = IMorphoCredit(address(morpho)).repaymentObligation(_id, borrower);
         (status,) = IMorphoCredit(address(morpho)).getRepaymentStatus(_id, borrower);
@@ -504,7 +504,7 @@ contract BaseTest is Test {
         pure
         returns (uint256)
     {
-        return endingBalance.wMulDown(penaltyRate.wTaylorCompounded(penaltyDuration));
+        return uint256(endingBalance).wMulDown(penaltyRate.wTaylorCompounded(penaltyDuration));
     }
 
     function _assertRepaymentStatus(Id _id, address borrower, RepaymentStatus expectedStatus) internal {
